trying to debug SubmodelComp
initial_mass is being passed/set to 1 in the print_comp and the traj
even though the correct value is being calculated in the exec_comp/
the explicit_comp. This happens even if set_input_defaults is used
to set the initial value to 154000. Also happens when the exec_comp
equation is modified to 'mass_initial = 175000'. I also tried using
an exec_comp between the initial_mass calculation and the traj to
separate the variables, but the issue still persisted.

Added a killer comp that stops the execution after the first timestep
in the trajectory runs, this makes it easier to compare what values
are being used, without them getting burried.