Added check for submodel bug fix in OpenMDAO
If the fix is not present, the old descent fuel as a function will
be used, if the fix is present (requires a specific dev version of
om) the descent fuel as a submodel will be used. Using this check
means that users who do not use shooting will not need the dev
version of OpenMDAO to get their tests to pass.