restored debugging output and updated test values
using fix_initial=True for altitude originally resulted in
EXIT: Error in step computation!
removing this caused the final state of the optimization to be
EXIT: Maximum Number of Iterations Exceeded.
this change significantly improved convergence for multiple cases,
and required updating the test values here, which were not originally
based on truth values anyways