fixed bugs related to specifying phase_info from the CLI
phase_info was being treated as a dict when it was passed to
run_aviary, added a preprocessor to run_level_1 that extracts
phase_info and phase_info_parameterization (if present) from
the file path that is specified through the CLI and then passes
them to run_aviary.
Also prevented the FileNotFoundError raised in get_model from
causing get_path to fail prematurely.